[incompat,VxWorks,->RELENG_2_0] #4157
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist/trunk@3016 8bdb766d-82ee-4cb2-b2de-c55db819e4d5

+++ src/lib/coil/vxworks/coil/UUID.cpp
@@ -116,8 +116,10 @@ namespace coil
 	  result += "-";
 	  result += StringToUUID<uint16_t>(_uuid.time_hi_version,4);
 	  result += "-";
-	  result += StringToUUID<uint8_t>(_uuid.clock_seq_low,2);
-	  result += StringToUUID<uint8_t>(_uuid.clock_seq_hi_variant,2);
+	  result += StringToUUID<uint16_t>((uint16_t)_uuid.clock_seq_low,2);
+	  result += StringToUUID<uint16_t>((uint16_t)_uuid.clock_seq_hi_variant,2);
+	  //result += StringToUUID<uint8_t>(_uuid.clock_seq_low,2);
+	  //result += StringToUUID<uint8_t>(_uuid.clock_seq_hi_variant,2);
 	  result += "-";
 	  result += StringToUUID<uint32_t>(_uuid.node_low,8);
 	  result += StringToUUID<uint16_t>(_uuid.node_high,4);
